Function strconv::parse_int

pub fn parse_int(s: &'static str, base: u8, bit_size: u8) -> Result<i64, NumError>

Interprets a string s in the given base (0, 2 to 36) and bit size (0 to 64) and returns the corresponding value i.

The string may begin with a leading sign: “+” or “-”.

If the base argument is 0, the true base is implied by the string’s prefix following the sign (if present): 2 for “0b”, 8 for “0” or “0o”, 16 for “0x”, and 10 otherwise. Also, for argument base 0 only, underscore characters are permitted as defined by the Go syntax for integer literals.

The bit_size argument specifies the integer type that the result must fit into. Bit sizes 0, 8, 16, 32, and 64 correspond to int, int8, int16, int32, and int64. If bit_size is below 0 or above 64, an error is returned.

The errors that parse_int returns have concrete type NumError and include err.num = s. If s is empty or contains invalid digits, err.err = NumErrorCause::InvalidSyntax and the returned value is 0; if the value corresponding to s cannot be represented by a signed integer of the given size, err.err = NumErrorCause::OutOfRangeSigned and the returned value is the maximum magnitude integer of the appropriate bit_size and sign.

Example

fn main() {
    const V32: &'static str = "-354634382";

    let got = strconv::parse_int(V32, 10, 32).unwrap();
    assert_eq!(got, -354634382i64);

    let _ = strconv::parse_int(V32, 16, 32).unwrap_err();

    const V64: &'static str = "-3546343826724305832";

    let got = strconv::parse_int(V64, 10, 64).unwrap();
    assert_eq!(got, -3546343826724305832);

    let _ = strconv::parse_int(V64, 16, 64).unwrap_err();
}